requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1

selenium==4.25.0
playwright==1.55.0
pandas==2.3.3
requests==2.32.3
boto3==1.35.49

openpyxl==3.1.5
pdfplumber==0.11.4
tabula-py==2.9.3
python-dotenv==1.0.1
psutil==5.9.8

